SOS: Fix wrong connection lifecycle handling
Quartz only uses one instance of the background job for all executions.
This is not reflected by the connection factory which will return the
closed connection from the first execution in every subsequent
execution. Fix this by resetting the factory on commit and rollback so
a new connection will be created.