STOCKS-101 Add system tests
Make sure the update-change test checks for changed dates. As long as
the table is empty, the update-change date is alwas now(). Fix the
initial change date by having a non-empty table.

Add a test to query for changed data. Make sure when requesting only
changed data that the server replies with only relevant changes