STOCKS-109: Add DB migration
Extend the DB model to store the creating user and device for every
record. To allow migration the existing devices' timestamps have to be
adjusted to model the instant creation of the first user and device
within the same transaction.

Increase the timeout for checked out connections because these
migrations are expensive.