STOCKS-116: Allow connection factory to recover
In a scenario where a background job's transaction fails during commit
or rollback the connection factory is not reset because the code after
the failing commit/rollback is not executed. Move connection factory
reset to the first operation. This can never fail and make sure the next
time the connection factory is called a fresh connection will be
returned regardless of the state of a former connection.