STOCKS-118: Show first event in outline fragment
This was a battle!

Simply reporting the list to be one element larger due to the header
view does not work entirely. When the fragment opens the recyclerview
scrolls to the bottom of the list without binding the last view, leaving
the chronologically first element empty. This is inacceptible.

Thus the header was moved out of the recyclerview again. This lead to
the insight that you cannot have all of this on android:

* A recyclerview loading only partial data
* A header which is scrolled as part of the recyclerview

Depending on the layout weird things happen. Items are shown twice while
being bound only once. During list animation items overlap in ugly ways.
The recyclerview scrolls into the middle of the list.

Thus the list is now embedded below the header and the header won't
disappear.