STOCKS-124: Enable young devices to act on data
Relax the foreign key constraints on the "initiates" fields to allow
devices created after an entity was created to act on that entity. This
is achieved by not applying the usual temporal foreign key constraint
but instead only verify that at the time of insertion of a new row the
device is valid. It doesn't have to exist for the entire lifetime of the
referencing row.