Issue #39 Remove hystrix
Hystrix needs a lot of resources (threads, RAM) and has offered little
benefit for this small project.

Even worse COMMITs take a very long time which hystrix should not be
configured to wait. If hystrix times out a request the REST API returns
a failure while the commit will eventually succeed. As there is no
reasonable way to abort a transaction during COMMIT, hystris is not a
good match for us.