Issue #42 Refactor spinner list suggestions
Introduce new type to represent a list with a suggestion for the UI.
Don't use it for conflict or food location as of now as the consequences
must be more carefully considered than I'm awake for now.

Fix bug resulting in no prefilled form data when adding food items with
no suggestable location.