Issue #42 Fix observable operators
zip() only fires downstream if all source observables have a new data
item. In contrast, combineLatest fires downstream as soon as any single
source observable has a new data item.

If the source observables change fire independent from each other (as in
the case of FoodItemListInteractorImpl) it is safe to use combineLatest.
In other cases (as in FoodByLocationListInteractorImpl) the
observing operation relies on being delivered a consistent state of
the data. Then using combineLatest is illegal as the race condition
could cause an inconsistent state to fire downstream or even crash the
operation as the consistency assumption is violated.