Issue #77 Fix rxjava handling
Subscribing an observable passed from android room makes the source
query reexecute. Thus when retrieving the list to handle click-events
has triggered the query to run twice, once for listing and once for the
click-event. This is inefficient as the list is already loaded.

Fix this by caching the list into a BehaviorSubject which is used for
click-events. Use the original observable for listing to have the
benefits of LiveData disposition.