Issue #47 Draft serial synchronisation
In order for synchronisation to be correct it must not be possible to
insert records with a transaction time smaller than the largest value
ever read from the updates view.

In case such a record existed, the client having read a larger value
from the updates view will never read that record, leading to an
inconsistent database on the client.

Since serializable transactions do not guarantee this, we have to
explain this additional constraint to postgresql. Every transaction
reads the current value of the transaction time clock and only proceeds
if its own transaction start time is larger. Finally it writes its
transaction start time into the transaction time clock before
committing.

As the tests show, this application behaviour makes postgresql enforce
the invariant for us.