Issue #99 Recipe food text autocompletion
Replace food spinner by autocompleting text view. This allows for faster
selection of the relevant food.

Android autocomplete textview doesn't allow for a strong connection of
the autocomplete suggestions and the text. This leads to validation
problems if a non-suggested text is entered. In this case, just ignore
the input and use the pre-initialised food of the ingredient / product.